# acceleration/api/openapi.yaml
openapi: 3.0.3
info:
  title: Model Router
  version: 0.2.0
  description: >
    Routes speech-to-text and text-to-speech traffic across providers and reports
    what it cost. Every path is scoped by modality, so the same provider serving
    two modalities is reported on separately. The caller is identified by a
    trusted X-Customer-Id header; real authentication is not part of this version.
servers:
  - url: http://localhost:8080
security:
  - CustomerId: []
paths:
...
  /v1/agents/sessions:
    get:
      operationId: listSessions
      summary: The sessions the calling customer is running
      responses:
        "200":
          description: The customer's sessions, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createSession
      summary: Join a call as a voice agent
      description: >
        The whole conversation runs here: the agent joins the call, transcribes what it
        hears, answers it and speaks back, all through the routers. The caller keeps the
        session id and watches the conversation over the events socket.

        It returns once the agent is in the call, so a session that comes back is one that
        is already listening. Tools declared here are the caller's own: the model asks for
        them over the events socket and waits for the caller to answer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: The agent is in the call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/agents/sessions/{id}:
    get:
      operationId: getSession
      summary: One session
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "200":
          description: The session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: closeSession
      summary: Leave the call and end the session
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "204":
          description: The agent has left
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/agents/sessions/{id}/events:
    get:
      operationId: watchSession
      summary: Watch the conversation and answer the model's tool calls
      description: >
        A WebSocket, which OpenAPI cannot describe past the upgrade. Frames are JSON
        objects carrying a `type` and the fields of that event.

        The server sends what the conversation did: `joined`, `heard`, `responding`,
        `response_delta`, `responded`, `spoke`, `turn`, `decision`, `delegated`,
        `task_settled`, `task_cancelled`, `tool_call`, `tool_ran`, `transferred`,
        `pressed`, `looked_up`, `backchannel`, `interrupted`, `overlap_decided`,
        `conversation_compacted`, `error` and `left`.

        A `decision` frame is one judgement the conversation made, carrying the same
        fields as a CallEvent. Together they are why the call went the way it did, and
        they are also written down, so a finished call replays them from
        `/v1/agents/calls/{id}/events`.

        Two frames are only sent when asked for, because they are far more frequent than
        the rest and most consumers want neither. `interim=true` adds `hearing`, which is
        a transcript revision as it arrives rather than a settled turn. `decisions=false`
        drops `decision`.

        The client sends `tool_result` to answer a `tool_call`, and `say`, `respond`,
        `interrupt`, `instructions` or `close` to act on the session. A `tool_call` is the
        only frame that must be answered: everything else is a report.
      parameters:
        - $ref: "#/components/parameters/SessionID"
        - name: interim
          in: query
          required: false
          description: Also send `hearing` frames, which are the words as they are revised.
          schema:
            type: boolean
            default: false
        - name: decisions
          in: query
          required: false
          description: Send `decision` frames.
          schema:
            type: boolean
            default: true
      responses:
        "101":
          description: The socket is open
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    CustomerId:
      type: apiKey
      in: header
      name: X-Customer-Id
      description: Trusted customer identifier. Every statistic is keyed by it.
  parameters:
...
    SessionID:
      name: id
      in: path
      required: true
      description: The session, as returned when it was created.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: The request was malformed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: The customer header is missing
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No such modality, provider or shortcut
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SessionState:
      type: string
      enum: [live, ended]
      description: Whether the agent is still in the call.
    SessionTool:
      type: object
      required: [name, description]
      description: >
        One of the caller's own functions. The model is offered it by name and description;
        running it is the caller's business, over the events socket.
      properties:
        name:
          type: string
        description:
          type: string
          description: >
            What the model is told the tool does, which is the whole of how it decides when
            to reach for one.
        parameters:
          type: object
          additionalProperties: true
          description: A JSON Schema object describing the arguments.
    SessionSkill:
      type: object
      required: [name, description, instructions]
      description: >
        A kind of work worth handing to the slower model. There is nothing behind a skill
        but a better model: what it declares is the instructions that model answers under.
      properties:
        name:
          type: string
        description:
          type: string
          description: The one line the fast model sees.
        instructions:
          type: string
          description: The full prompt, which only the subagent sees.
        deadline_ms:
          type: integer
          format: int64
          description: How long the work may run before it is abandoned. Zero is the default.
    SessionMemory:
      type: object
      description: >
        Who the session's memories are about. Without a user id nothing is recalled or
        stored, which is the case for a call with nobody identified on it.
      properties:
        user_id:
          type: string
          description: Who the memories belong to. Empty means the customer.
        app_id:
          type: string
          description: Separates two deployments sharing one memory account.
        filter:
          type: object
          additionalProperties:
            type: string
          description: >
            The caller's own labels, which narrow recall further. They cannot widen it: a
            filter is applied alongside the user id, never instead of it.
    SessionPhone:
      type: object
      required: [number]
      description: The number the session acts from, which is what turns transferring on.
      properties:
        number:
          type: string
          description: One of the customer's own numbers, written as +15551234567.
        vendor:
          type: string
          description: Who carries an outbound leg.
        vendor_call_id:
          type: string
          description: >
            The outbound leg, set for a call the agent placed. Without one the agent has no
            keypad to press at.
...
    CreateSessionRequest:
      type: object
      properties:
        call_id:
          type: string
          description: The call to join. Required unless the session is text.
        text:
          type: boolean
          default: false
          description: >
            Hold the conversation in writing rather than on a call. Nothing is transcribed
            and nothing is spoken, so no call is joined and neither speech target is used.
            Everything between hearing and answering is unchanged: a text session has the
            same skills, knowledge and tools a call would have had, and its replies arrive
            as response_delta and responded events on the session's socket.
        config_id:
          type: string
          description: >
            An agent config to start from. Everything else in this request overrides what
            the config says, so a caller can reuse a configuration and still change one
            thing about this call.
        call_type:
          type: string
          default: default
        user_id:
          type: string
          description: Who the agent joins the call as.
          default: vision-agent
        user_name:
          type: string
          default: Vision Agent
        agent_id:
          type: string
          description: Keys transcripts and statistics. Empty means the call id.
        instructions:
          type: string
        greeting:
          type: string
          description: >
            Said on joining without going through the model. Empty means the agent waits to
            be spoken to.
        navigating:
          type: boolean
          default: false
          description: >
            The agent placed this call, so let recordings finish and answer their menus.
        llm:
          type: string
          default: llm-fast
          description: A provider/model or a capability shortcut.
        stt:
          type: string
          default: en-low-latency
        tts:
          type: string
          default: en-low-latency
        subagent:
          type: string
          description: >
            The model that does the thinking. Empty means the voice model answers
            everything itself, and skills mean nothing.
        voice:
          type: string
          description: Provider-specific voice id.
        languages:
          type: array
          items:
            type: string
          description: Language hints, which narrow the candidates in every modality.
        keyterms:
          type: array
          items:
            type: string
          description: >
            Business-specific words the transcriber would otherwise get wrong. Up to 100
            terms, and providers that cannot be told about vocabulary ignore them.
        max_tokens:
          type: integer
        tasks:
          type: integer
          description: How much delegated work may run at once.
        sandbox:
          type: string
          enum: [daytona]
          description: >
            Where the subagent may run code it writes. Only the subagent is offered it:
            running code takes seconds, and the model holding the conversation has none to
            spare. Omit it and the subagent works everything out in its head.
        backchannel:
          type: boolean
          default: false
          description: Murmur while a participant is still talking, the way a person does.
        min_confidence:
          type: number
          format: double
          description: >
            How sure the transcriber must be before the agent answers rather than checks
            what was meant.
        skills:
          type: array
          items:
            $ref: "#/components/schemas/SessionSkill"
          description: Omit for the built-in set of think, recall and explain.
        skill_names:
          type: array
          items:
            type: string
          description: >
            Skills to look up rather than spell out: the customer's own, or one of the
            built-in think, recall and explain. Ignored when skills are given in full,
            and a name nothing defines is refused rather than dropped.
        tools:
          type: array
          items:
            $ref: "#/components/schemas/SessionTool"
        tool_timeout_ms:
          type: integer
          description: How long the model waits for a tool result. Zero is the default.
        tags:
          type: object
          additionalProperties:
            type: string
          description: Cost labels, carried onto every request the session makes.
        memory:
          $ref: "#/components/schemas/SessionMemory"
        phone:
          $ref: "#/components/schemas/SessionPhone"
    Session:
      type: object
      required: [id, call_id, call_type, user_id, agent_id, state, created_at]
      properties:
        id:
          type: string
        call_id:
          type: string
          description: Empty for a text session, which joins no call.
        text:
          type: boolean
          description: The conversation is held in writing rather than on a call.
        call_type:
          type: string
        user_id:
          type: string
        agent_id:
          type: string
        state:
          $ref: "#/components/schemas/SessionState"
        created_at:
          type: string
          format: date-time
        llm:
          type: string
          description: The provider and model answering, once routing has picked one.
        tts:
          type: string
          description: The provider and model speaking.
        instructions:
          type: string
...